pyproject.toml
[project]
name = "label-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "transformers @ git+https://github.com/huggingface/transformers.git",
    "torch>=2.0.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "torchvision>=0.24.1",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "httpx>=0.28.1",
    "boto3>=1.34.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[dependency-groups]
dev = [
    "ruff>=0.14.5",
]
